build: Update deployment target for all platforms
It appears that you have updated the deployment target for your TVOS project to version 12.0 in your Xcode project settings. This means that your app will only run on devices or simulators with tvOS 12.0 or later installed. If you want to support older versions of tvOS, you should adjust the deployment target accordingly.

   Additionally, you have also set the IPHONEOS_DEPLOYMENT_TARGET to 12.0 for your iOS project. This means that your app will only run on devices or simulators with iOS 12.0 or later installed. If you want to support older versions of iOS, you should adjust the deployment target accordingly.

   Keep in mind that updating the deployment target may affect the compatibility of your app with certain devices and features. Always test your app thoroughly before releasing it to ensure that it works as intended on all supported platform