refactor(Punycode.xcodeproj): Renamed Punycode-visionOS XCScheme for testing, profiling, and archiving
This commit adds a new XCScheme file for the Punycode project named "Punycode-visionOS". The scheme supports building for testing, running, profiling, archiving, and analyzing. It also includes code coverage for specified targets and testables for the Punycode-visionOSTests.xctest. This change will facilitate better testing and debugging of the Punycode project in Xcode.