feat(fastlane): Add Fastlane plugin versioning and update Pluginfile
Added fastlane-plugin-versioning to the project's Fastlane Pluginfile.
This will allow for automatic versioning of the app when building with Fastlane.

refactor(run.sh): Update run script to support "clean xcode" command

Updated the run script to include a new "clean xcode" command that cleans the Xcode project and derived data, erases simulators, and removes the DerivedData folder.
This command is now included in the list of available commands when running the script with fzf.

build(fastlane): Update Fastlane commands to support multiple platforms

Updated the Fastlane commands for Carthage to update for all supported platforms (macOS, iOS, tvOS, watchOS, and visionOS).
This will ensure that the correct Carthage files are updated for each platform when building the app with Fastlane.